this the second step
