+++ package.json
@@ -25,7 +25,7 @@
     "swagger": "swagger-repo",
     "test": "swagger-repo validate",
     "start": "gulp serve",
-    "yaml-parse": "node ./spec/yamltest.js",
+    "process-yaml": "node ./scripts/process-yaml.js",
     "deploy-branch": "node ./scripts/deploy-branch.js"
   }
 }
