+++ scripts/process-yaml.js
@@ -1,0 +1,51 @@
+const yaml = require('js-yaml');
+const fs = require('fs');
+const { Resolver }  = require('@stoplight/json-ref-resolver');
+const { join } = require('@stoplight/path')
+
+try {
+    const myFunction = async (sourceFilePath, destinationFilePath) => {
+        // Get document, or throw exception on error
+        console.log("-----------------------------------------------------------");
+        console.log("process sourceFilePath=" + sourceFilePath);
+        const doc = yaml.load(fs.readFileSync(sourceFilePath, 'utf8'));
+        const baseFilePath  = "./spec/translation_quality";
+        const resolver = new Resolver({
+            // resolvers can do anything, so long as they define an async read function that resolves to a value
+            resolvers: {
+                // this resolver will be invoked for refs with the file protocol
+                baseURI: baseFilePath,
+                file: {
+                    resolve(ref) {
+                        console.log("before resolve ref");
+                        console.log(ref);
+                        const docUri = join(baseFilePath, ref.toString());
+                        return yaml.load(fs.readFileSync(docUri, 'utf8'));
+                    }
+                }
+            }
+        });
+        console.log("-----------------------------------------------------------");
+        console.log("before resolver.resolve(..)");
+        // const resolved = await resolver.resolve(doc, { jsonPointer: "#/root" });
+        const resolved = await resolver.resolve(doc, { jsonPointer: "#/x-paths" });
+        console.log("after resolver.resolve(..)");
+        console.log("-----------------------------------------------------------");
+        console.log(resolved);
+        console.log("before resolved.result");
+        var destinationJson = {"x-path": resolved.result}
+        // fs.writeFileSync(destinationFilePath, yaml.dump(resolved.result));
+        fs.writeFileSync(destinationFilePath, yaml.dump(destinationJson));
+        console.log("after resolved.result");
+        console.log("processed sourceFilePath=" + sourceFilePath);
+        console.log("-----------------------------------------------------------");
+    }
+    myFunction("./spec/translation_quality/dictionaries.yaml", "./spec/translation_quality/dictionaries_result.yaml");
+    myFunction("./spec/translation_quality/profiles.yaml", "./spec/translation_quality/profiles_result.yaml");
+    myFunction("./spec/translation_quality/checks/accounts.yaml", "./spec/translation_quality/checks/accounts_result.yaml");
+    myFunction("./spec/translation_quality/check_type_severity_level.yaml", "./spec/translation_quality/check_type_severity_level_result.yaml");
+    myFunction("./spec/translation_quality/account_settings.yaml", "./spec/translation_quality/account_settings_result.yaml");
+    myFunction("./spec/translation_quality/tqc_check.yaml", "./spec/translation_quality/tqc_check_result.yaml");
+} catch (e) {
+    console.log(e);
+}
