--- spec/example1.yml
@@ -1,36 +1,0 @@
-components:
-  schemas:
-    dictionary_checkTypes:
-      get:
-        description: >-
-          Get the quality check types that are available.
-        summary: Get quality check types
-        tags:
-          - Translation quality checks
-        operationId: getTranslationQualityCheckTypes
-        consumes:
-          - application/json
-        produces:
-          - application/json
-
-        responses:
-          r200:
-            description: OK
-            content:
-#              application/json:
-              applicationjson:
-                schema:
-#                  $ref: '#/components/schemas/DictionaryCheckTypesResponse'
-                example:
-                  $ref: '#/components/examples/dictionary_severity_levels_example'
-
-  examples:
-    dictionary_severity_levels_example:
-      response:
-        code: SUCCESS
-        data:
-#            - $ref: '#/components/examples/myTest'
-          $ref: './spec/example2.yml#/components/examples/myTest'
-
-    myTest:
-      prop1: 300
