--- spec/example2.yml
@@ -1,5 +1,0 @@
-components:
-  examples:
-    myTest:
-      card: 100
-      type: "debit"
