--- spec/example_result.yml
@@ -1,22 +1,0 @@
-get:
-  description: Get the quality check types that are available.
-  summary: Get quality check types
-  tags:
-    - Translation quality checks
-  operationId: getTranslationQualityCheckTypes
-  consumes:
-    - application/json
-  produces:
-    - application/json
-  responses:
-    r200:
-      description: OK
-      content:
-        applicationjson:
-          schema: null
-          example:
-            response:
-              code: SUCCESS
-              data:
-                card: 100
-                type: debit
