+++ spec/translation_quality/dictionaries.yaml
@@ -1,5 +1,4 @@
 x-paths:
-
   dictionary_checkTypes:
     get:
       description: >-
